+++ database/src/desktopMain/kotlin/ru/pavlig43/database/data/storage/StorageProduct.kt
@@ -3,18 +3,18 @@ package ru.pavlig43.database.data.storage
 data class StorageProduct(
     val productId: Int,
     val productName: String,
-    val balanceBeforeStart: Int,
-    val incoming: Int,
-    val outgoing: Int,
-    val balanceOnEnd: Int,
+    val balanceBeforeStart: Long,
+    val incoming: Long,
+    val outgoing: Long,
+    val balanceOnEnd: Long,
 
     val batches: List<StorageBatch>
 )
 data class StorageBatch(
     val batchId: Int,
     val batchName: String,
-    val balanceBeforeStart: Int,
-    val incoming: Int,
-    val outgoing: Int,
-    val balanceOnEnd: Int,
+    val balanceBeforeStart: Long,
+    val incoming: Long,
+    val outgoing: Long,
+    val balanceOnEnd: Long,
 )
