+++ features/form/transaction/src/desktopMain/kotlin/ru/pavlig43/transaction/internal/update/tabs/component/opzs/ingredients/FillIngredientsRepository.kt
@@ -8,7 +8,7 @@ internal class FillIngredientsRepository(
 ) {
     private val compositionDao = db.compositionDao
 
-    suspend fun getIngredientsFromComposition(productId: Int,transactionId: Int,countPf: Int): Result<List<IngredientBD>> {
+    suspend fun getIngredientsFromComposition(productId: Int,transactionId: Int,countPf: Long): Result<List<IngredientBD>> {
         return runCatching {
             compositionDao.getIngredientsFromComposition(productId,transactionId,countPf)
         }
