+++ features/table/immutable/src/desktopMain/kotlin/ru/pavlig43/immutable/internal/component/items/expense/ExpenseTableComponent.kt
@@ -34,7 +34,7 @@ private fun MainExpenseBD.toUi(): ExpenseTableUi {
     return ExpenseTableUi(
         composeId = expense.id,
         expenseType = expense.expenseType,
-        amount = DecimalData2(expense.amount),
+        amount = DecimalData2(expense.amount.toLong()),
         expenseDateTime = expense.expenseDateTime,
         comment = expense.comment,
         transactionId = transaction?.id
